Znakomstvo s Python/lesson_3-Rassylaiem_imieily/requirements.txt
python-dotenv == 0.11.0

